m68k: mac: fix y2036 overflow for RTC
The real-time clock on m68k mac systems uses an unsigned 32-bit
value starting in 1904, which overflows in 2036, about two years
earlier than everyone else.

I suppose we don't really need to honor RTC timestamps before 1970,
as they would not be correct anyway, so this introduces a m68k
mac specific RTC windowing that interprets all dates between 1904
and 1969 as times past the 2036 overflow, by adding a whole epoch
during read-out. When writing to the RTC, no change is needed
since the lower 32 bits of the timestamp are unaffected.

To actually make it work beyond 2038, I'm also changing the types
from 'long' to 'time64_t'. This has the benefit of removing one
of the last callers of the deprecated mktime() interface.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>