[SUBMITTED 20180713] [v2, RESEND] cachefiles: avoid deprecated get_seconds()
get_seconds() returns an unsigned long can overflow on some architectures
and is deprecated because of that. In cachefs, we cast that number to
a a 32-bit integer, which will overflow in year 2106 on all architectures.

As confirmed by David Howells, the overflow probably isn't harmful
in the end, since the timestamps are only used to make the file names
unique, but they don't strictly have to be in monotonically increasing
order since the files only exist in order to be deleted as quickly
as possible.

Moving to ktime_get_real_seconds() avoids the deprecated interface.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
v2: only change to ktime_get_real_seconds() without printing the
    extra leading digits

    Originally submitted this on June 20, but got no reply.
    Resending this in case it got lost somewhere on the way.
    David, can you pick this up?