shmem: use monotonic time for i_generation
get_seconds() is deprecated because it will lead to a 32-bit overflow
in 2038 or 2106. For the tmpfs i_generation update, we can simply use
ktime_get_seconds() instead, which uses monotonic time and won't
overflow.

The monotonic timestamps can repeat after a reboot, but as tmpfs is
not persistent, that should not be a problem.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>