[SUBMITTED 20180713] [v2, RESEND] afs: avoid deprecated get_seconds()
get_seconds() has a limited range on 32-bit architectures and is
deprecated because of that. While AFS uses the same limits for
its inode timestamps on the wire protocol, let's just use the
simpler current_time() as we do for other file systems.

This will still zero out the 'tv_nsec' field of the timestamps
internally.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
Originally submitted this on June 20, but got no reply.
Resending this in case it got lost somewhere on the way.
David, can you pick up both patches?