y2038: provide updated getitimer/setitimer syscalls
I had originally planned to not provide getitimer/setitimer system calls
with time_t arguments, given that the posix timer interfaces are a far
superior replacement.

However, I found during testing that it's not possible to
emulate the getitimer/setitimer interface from user space using
timer_gettime/timer_settime due to small differences in the behavior
across execve() and the way that the IDs are managed dynamically.

The options I considered instead are:

1. Not changing the system call interface at all, but requiring the C
   library to convert between the application-facing types (with 64-bit
   time_t) and the kernel-facing 32-bit types. This can work because
   all time arguments are relative times that do not suffer from the
   y2038 problem. It would however mean truncating intervals to INT_MAX
   seconds, which could result in expected behavior.

2. Using one of the flag bits in timer_gettime/timer_settime to make
   those behave like getitimer/setitimer, but taking an itimerspec argument
   with 64-bit time_t. This could even incorporate the timerfd variants,
   reducing the number of syscalls from six to two here.

3. Adding a flag bit to the 'which' argument that decides on the format
   of the other arguments, either using the 32-bit version
   of timeval/itimerval components or the 64-bit version of
   timespec/itimerspec

4. Adding a pair of new system calls with only 64-bit time_t based
   arguments, and the best possible interface.

This implements the last of those four, as it seems to be the least
ugly solution. I completely remove the itimerval usage from the kernel,
passing around itimerspec64 and timespec64 structures inside the kernel,
like we do with the rest of the time handling these days.  On the user
interface side, I use the same prototype as the timer and timerfd system
calls for consistency. The flags word is not used yet, aside from being
required to be zero.

Using a timespec based structure instead of timeval means we can avoid
adding a 64-bit version of timeval to the kernel uapi headers, which
seems appropriate given that all timeval based interfaces have been
deprecated by newer alternatives, and we don't want any new timeval
interfaces to creep in. It also means that we get the native resolution
of the timekeeping code, rather than having to divide by 1000 in the
kernel. However, it also means that user space now does have to do this
division in order to implement the traditional interfaces.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>