ARM: disable EFI stub for clang
The EFI stub relies on building with gcc -mno-single-pic-base, which
is not supported by clang, causing a build failure.

Ard mentioned that he had some ideas how to work around this
problem in the past, but for now, we might disable the
feature entirely until a fix has been found.

Link: https://lkml.org/lkml/2017/5/10/538
Cc: Bernhard Rosenkränzer <Bernhard.Rosenkranzer@linaro.org>
Cc: Matthias Kaehlcke <mka@chromium.org>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>