ARM: fix big-endian build with clang
Trying to build a big-endian kernel with clang fails because both the
assembler and linker called by it fall back to little-endian mode without
the -EB command line argument.

This is a bug in clang, which may be fixed in the clang-8 release, but
we still need to work around older releases, so let's add the arguments
manually.

Link: https://bugs.llvm.org/show_bug.cgi?id=38770
Signed-off-by: Arnd Bergmann <arnd@arndb.de>