ARM: update -march= flag selection
Clang has problems with some of the -march= fallbacks, which gives
incorrect build results.

We don't actually need the fallbacks for armv6/7 any more, because gcc-4.6
is now the minimum supported compiler version and armv7 support has been
available since gcc-4.3.

For armv6k, clang sets the CPU to arm1176j-s, which the GNU assembler
does not understand, so we have to override it to the almost identical
arm1176jz-s.

For ARMv4, we need to keep the fallback to ARMv4T (with the --fix-v4bx
linker option) to still keep newer compilers working that remove support
for the original ARMV4.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>