ARM: use explicit 'it' instructions in cmpxchg
clang sometimes makes a mistake with calculating branch lengths in
the presence of 'it' instructions added by the assembler on a thumb2
kernel when using the -mimplicit-it=always flag. This results in a build
failure like:

/tmp/iscsi_target-5348fd.s: Assembler messages:
/tmp/iscsi_target-5348fd.s:8071: Error: branch out of range

As a workaround, use explicit instructions in cmpxchg and related
functions that are known to have triggered this at some point.

Link: https://bugs.llvm.org/show_bug.cgi?id=31805
Signed-off-by: Arnd Bergmann <arnd@arndb.de>